backend/pyproject.toml
[project]
name = "intelligent-doc-qa"
version = "0.1.0"
description = "Self-reflective LangGraph RAG system for document question answering"
requires-python = ">=3.10"
authors = [{ name = "Vivek Pandey" }]
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "langchain>=0.2.3",
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.0",
    "langgraph>=0.0.44",
    "chromadb>=0.5.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.4",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.16",
    "sqlalchemy>=2.0.28",
    "psycopg[binary]>=3.1.18",
    "duckdb>=1.0.0",
    "pypdf>=4.2.0",
    "python-docx>=1.1.0",
    "unstructured[all-docs]>=0.12.5",
    "langchain-text-splitters>=0.2.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.6.0",
    "nltk>=3.8.1",
    "spacy>=3.7.4",
    "tiktoken>=0.7.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "ragas>=0.1.17",
    "jupyter>=1.0.0",
    "httpx>=0.27.0",
    "requests>=2.32.0",
    "typer>=0.12.3",
    "rich>=13.7.1",
    "langchain-ollama>=0.2.0",
    "langchainhub>=0.1.15",
    "einops>=0.7.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "mypy>=1.9.0",
    "ruff>=0.4.3",
    "black>=24.4.0"
]

[tool.uv]
package = "intelligent-doc-qa"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "DTZ", "PT", "RET", "SIM", "PL"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
addopts = "-q --disable-warnings"
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
